ml-worker/requirements.txt
redis>=5.0
google-genai>=1.0
pyannote.audio>=3.3
torch>=2.0
transformers>=4.40
pydantic>=2.0
pydantic-settings>=2.0
